Исправляет ошибки(5.12)
